+++ mycodo/functions/regulate_ph_ec.py
@@ -34,13 +34,7 @@ from mycodo.utils.constraints_pass import constraints_pass_positive_value
 from mycodo.utils.database import db_retrieve_table_daemon
 from mycodo.utils.send_data import send_email
 
-measurements_dict = {
-    0: {
-        'measurement': '',
-        'unit': '',
-        'name': ''
-    }
-}
+measurements_dict = {0: {}}
 
 FUNCTION_INFORMATION = {
     'function_name_unique': 'REGULATE_PH_EC',
@@ -48,14 +42,10 @@ FUNCTION_INFORMATION = {
     'measurements_dict': measurements_dict,
     'enable_channel_unit_select': False,
 
-    'message': 'This function regulates pH with 2 pumps (acid and base solutions) and electrical conductivity (EC) with up to 4 pumps (nutrient solutions A, B, C, and D). Set only the nutrient solution outputs you want to use. Any outputs not set will not dispense when EC is being adjusted, allowing as few as 1 pump or as many as 4 pumps. If an e-mail address (or multiple addresses separated by commas) is entered into the E-Mail Notification field, a notification e-mail will be sent if 1) pH is outside the set danger range, 2) EC is too high and water needs to be added to the reservoir, or 3) a measurement could not be found in the database for the specific Max Age. Each e-mail notification type has its own timer that prevents e-mail spam, and will only allow sending for each notification type every set E-Mail Timer Duration. After this duration, the timer will automatically reset to allow new notifications to be sent. You may also manually reset e-mail timers at any time with the Custom Commands, below.',
+    'message': 'This function regulates pH with 2 pumps (acid and base solutions) and electrical conductivity (EC) with up to 4 pumps (nutrient solutions A, B, C, and D). Set only the nutrient solution outputs you want to use. Any outputs not set will not dispense when EC is being adjusted, allowing as few as 1 pump or as many as 4 pumps. Outputs can be instructed to turn on for durations (seconds) or volumes (ml). Set each Output Type to the correct type for each selected Output Channel (only select on/off Output Channels for durations and volume Output Channels for volumes). If an e-mail address (or multiple addresses separated by commas) is entered into the E-Mail Notification field, a notification e-mail will be sent if 1) pH is outside the set danger range, 2) EC is too high and water needs to be added to the reservoir, or 3) a measurement could not be found in the database for the specific Max Age. Each e-mail notification type has its own timer that prevents e-mail spam, and will only allow sending for each notification type every set E-Mail Timer Duration. After this duration, the timer will automatically reset to allow new notifications to be sent. You may also manually reset e-mail timers at any time with the Custom Commands, below.',
 
-    'options_enabled': [
-        'custom_options'
-    ],
-    'options_disabled': [
-        'measurements_select'
-    ],
+    'options_enabled': ['custom_options'],
+    'options_disabled': ['measurements_select'],
 
     'custom_commands': [
         {
@@ -99,7 +89,24 @@ FUNCTION_INFORMATION = {
             'phrase': lazy_gettext('The duration (seconds) between measurements or actions')
         },
         {
-            'type': 'new_line'
+            'id': 'start_offset',
+            'type': 'integer',
+            'default_value': 10,
+            'required': True,
+            'name': 'Start Offset',
+            'phrase': 'The duration (seconds) to wait before the first operation'
+        },
+        {
+            'id': 'period_status',
+            'type': 'integer',
+            'default_value': 60,
+            'required': True,
+            'name': 'Status Period (seconds)',
+            'phrase': 'The duration (seconds) to update the Function status on the UI'
+        },
+        {
+            'type': 'message',
+            'default_value': "Measurement Options"
         },
         {
             'id': 'select_measurement_ph',
@@ -143,7 +150,8 @@ FUNCTION_INFORMATION = {
             'phrase': lazy_gettext('The maximum age (seconds) of the measurement to use')
         },
         {
-            'type': 'new_line'
+            'type': 'message',
+            'default_value': "Output Options"
         },
         {
             'id': 'output_ph_raise',
@@ -168,19 +176,31 @@ FUNCTION_INFORMATION = {
             'phrase': 'Select an output to lower the pH'
         },
         {
-            'id': 'output_ph_duration_sec',
+            'id': 'output_ph_type',
+            'type': 'select',
+            'default_value': 'duration_sec',
+            'required': True,
+            'options_select': [
+                ('duration_sec', 'Duration (seconds)'),
+                ('volume_ml', 'Volume (ml)')
+            ],
+            'name': 'pH Output Type',
+            'phrase': 'Select the output type for the selected Output Channel'
+        },
+        {
+            'id': 'output_ph_amount',
             'type': 'float',
             'default_value': 2.0,
             'required': True,
             'constraints_pass': constraints_pass_positive_value,
-            'name': 'pH Output Duration (Seconds)',
-            'phrase': 'The number of seconds to run the pH dosing pumps'
+            'name': 'pH Output Amount',
+            'phrase': 'The amount to send to the pH dosing pumps (duration or volume)'
         },
         {
             'type': 'new_line'
         },
         {
-            'id': 'output_ec_nutrient_a',
+            'id': 'output_ec_a',
             'type': 'select_channel',
             'default_value': '',
             'required': True,
@@ -191,19 +211,31 @@ FUNCTION_INFORMATION = {
             'phrase': 'Select an output to dose nutrient A'
         },
         {
-            'id': 'output_ec_a_duration_sec',
+            'id': 'output_ec_a_type',
+            'type': 'select',
+            'default_value': 'duration_sec',
+            'required': True,
+            'options_select': [
+                ('duration_sec', 'Duration (seconds)'),
+                ('volume_ml', 'Volume (ml)')
+            ],
+            'name': 'Nutrient A Output Type',
+            'phrase': 'Select the output type for the selected Output Channel'
+        },
+        {
+            'id': 'output_ec_a_amount',
             'type': 'float',
             'default_value': 2.0,
             'required': True,
             'constraints_pass': constraints_pass_positive_value,
-            'name': 'Nutrient A Output Duration (Seconds)',
-            'phrase': 'The number of seconds to run the Nutrient A dosing pump'
+            'name': 'Nutrient A Output Amount',
+            'phrase': 'The amount to send to the Nutrient A dosing pump (duration or volume)'
         },
         {
             'type': 'new_line'
         },
         {
-            'id': 'output_ec_nutrient_b',
+            'id': 'output_ec_b',
             'type': 'select_channel',
             'default_value': '',
             'required': True,
@@ -214,19 +246,31 @@ FUNCTION_INFORMATION = {
             'phrase': 'Select an output to dose nutrient B'
         },
         {
-            'id': 'output_ec_b_duration_sec',
+            'id': 'output_ec_b_type',
+            'type': 'select',
+            'default_value': 'duration_sec',
+            'required': True,
+            'options_select': [
+                ('duration_sec', 'Duration (seconds)'),
+                ('volume_ml', 'Volume (ml)')
+            ],
+            'name': 'Nutrient B Output Type',
+            'phrase': 'Select the output type for the selected Output Channel'
+        },
+        {
+            'id': 'output_ec_b_amount',
             'type': 'float',
             'default_value': 2.0,
             'required': True,
             'constraints_pass': constraints_pass_positive_value,
-            'name': 'Nutrient B Output Duration (Seconds)',
-            'phrase': 'The number of seconds to run the Nutrient B dosing pump'
+            'name': 'Nutrient B Output Amount',
+            'phrase': 'The amount to send to the Nutrient B dosing pump (duration or volume)'
         },
         {
             'type': 'new_line'
         },
         {
-            'id': 'output_ec_nutrient_c',
+            'id': 'output_ec_c',
             'type': 'select_channel',
             'default_value': '',
             'required': True,
@@ -237,19 +281,31 @@ FUNCTION_INFORMATION = {
             'phrase': 'Select an output to dose nutrient C'
         },
         {
-            'id': 'output_ec_c_duration_sec',
+            'id': 'output_ec_c_type',
+            'type': 'select',
+            'default_value': 'duration_sec',
+            'required': True,
+            'options_select': [
+                ('duration_sec', 'Duration (seconds)'),
+                ('volume_ml', 'Volume (ml)')
+            ],
+            'name': 'Nutrient C Output Type',
+            'phrase': 'Select the output type for the selected Output Channel'
+        },
+        {
+            'id': 'output_ec_c_amount',
             'type': 'float',
             'default_value': 2.0,
             'required': True,
             'constraints_pass': constraints_pass_positive_value,
-            'name': 'Nutrient C Output Duration (Seconds)',
-            'phrase': 'The number of seconds to run the Nutrient C dosing pump'
+            'name': 'Nutrient C Output Amount',
+            'phrase': 'The amount to send to the Nutrient C dosing pump (duration or volume)'
         },
         {
             'type': 'new_line'
         },
         {
-            'id': 'output_ec_nutrient_d',
+            'id': 'output_ec_d',
             'type': 'select_channel',
             'default_value': '',
             'required': True,
@@ -260,16 +316,29 @@ FUNCTION_INFORMATION = {
             'phrase': 'Select an output to dose nutrient D'
         },
         {
-            'id': 'output_ec_d_duration_sec',
+            'id': 'output_ec_d_type',
+            'type': 'select',
+            'default_value': 'duration_sec',
+            'required': True,
+            'options_select': [
+                ('duration_sec', 'Duration (seconds)'),
+                ('volume_ml', 'Volume (ml)')
+            ],
+            'name': 'Nutrient D Output Type',
+            'phrase': 'Select the output type for the selected Output Channel'
+        },
+        {
+            'id': 'output_ec_d_amount',
             'type': 'float',
             'default_value': 2.0,
             'required': True,
             'constraints_pass': constraints_pass_positive_value,
-            'name': 'Nutrient D Output Duration (Seconds)',
-            'phrase': 'The number of seconds to run the Nutrient D dosing pump'
+            'name': 'Nutrient D Output Amount',
+            'phrase': 'The amount to send to the Nutrient D dosing pump (duration or volume)'
         },
         {
-            'type': 'new_line'
+            'type': 'message',
+            'default_value': "Setpoint Options"
         },
         {
             'id': 'setpoint_ph',
@@ -332,7 +401,8 @@ FUNCTION_INFORMATION = {
             'phrase': 'This low pH value for the danger range'
         },
         {
-            'type': 'new_line'
+            'type': 'message',
+            'default_value': "Alert Notification Options"
         },
         {
             'id': 'email_notification',
@@ -350,7 +420,7 @@ FUNCTION_INFORMATION = {
             'constraints_pass': constraints_pass_positive_value,
             'name': 'E-Mail Timer Duration (Hours)',
             'phrase': 'How long to wait between sending e-mail notifications'
-        },
+        }
     ]
 }
 
@@ -380,23 +450,28 @@ class CustomModule(AbstractFunction):
         self.output_ph_raise_channel_id = None
         self.output_ph_lower_device_id = None
         self.output_ph_lower_channel_id = None
-        self.output_ph_duration_sec = None
+        self.output_ph_type = None
+        self.output_ph_amount = None
 
-        self.output_ec_nutrient_a_device_id = None
-        self.output_ec_nutrient_a_channel_id = None
-        self.output_ec_a_duration_sec = None
+        self.output_ec_a_device_id = None
+        self.output_ec_a_channel_id = None
+        self.output_ec_a_type = None
+        self.output_ec_a_amount = None
 
-        self.output_ec_nutrient_b_device_id = None
-        self.output_ec_nutrient_b_channel_id = None
-        self.output_ec_b_duration_sec = None
+        self.output_ec_b_device_id = None
+        self.output_ec_b_channel_id = None
+        self.output_ec_b_type = None
+        self.output_ec_b_amount = None
 
-        self.output_ec_nutrient_c_device_id = None
-        self.output_ec_nutrient_c_channel_id = None
-        self.output_ec_c_duration_sec = None
+        self.output_ec_c_device_id = None
+        self.output_ec_c_channel_id = None
+        self.output_ec_c_type = None
+        self.output_ec_c_amount = None
 
-        self.output_ec_nutrient_d_device_id = None
-        self.output_ec_nutrient_d_channel_id = None
-        self.output_ec_d_duration_sec = None
+        self.output_ec_d_device_id = None
+        self.output_ec_d_channel_id = None
+        self.output_ec_d_type = None
+        self.output_ec_d_amount = None
 
         self.setpoint_ph = None
         self.hysteresis_ph = None
@@ -410,6 +485,7 @@ class CustomModule(AbstractFunction):
 
         self.range_ph = None
         self.range_ec = None
+        self.ph_type = None
 
         self.output_ph_raise_channel = None
         self.output_ph_lower_channel = None
@@ -425,6 +501,15 @@ class CustomModule(AbstractFunction):
             "notify_none": 0
         }
 
+        self.output_types = {
+            'duration_sec': 'sec',
+            'volume_ml': 'ml'
+        }
+
+        self.list_doses = []
+        self.ratio_letters = []
+        self.ratio_numbers = []
+
         # Set custom options
         custom_function = db_retrieve_table_daemon(
             CustomController, unique_id=self.unique_id)
@@ -435,6 +520,8 @@ class CustomModule(AbstractFunction):
             self.try_initialize()
 
     def initialize(self):
+        self.timer_loop = time.time() + self.start_offset
+
         self.range_ph = (self.setpoint_ph - self.hysteresis_ph, self.setpoint_ph + self.hysteresis_ph)
         self.range_ec = (self.setpoint_ec - self.hysteresis_ec, self.setpoint_ec + self.hysteresis_ec)
 
@@ -444,16 +531,53 @@ class CustomModule(AbstractFunction):
             self.output_ph_lower_channel_id)
 
         self.output_ec_a_channel = self.get_output_channel_from_channel_id(
-            self.output_ec_nutrient_a_channel_id)
+            self.output_ec_a_channel_id)
         self.output_ec_b_channel = self.get_output_channel_from_channel_id(
-            self.output_ec_nutrient_b_channel_id)
+            self.output_ec_b_channel_id)
         self.output_ec_c_channel = self.get_output_channel_from_channel_id(
-            self.output_ec_nutrient_c_channel_id)
+            self.output_ec_c_channel_id)
         self.output_ec_d_channel = self.get_output_channel_from_channel_id(
-            self.output_ec_nutrient_d_channel_id)
+            self.output_ec_d_channel_id)
+
+        self.ph_type = 'sec'
+        if self.output_ph_type == "volume_ml":
+            self.ph_type = 'vol'
+
+        self.ec_a_type = 'sec'
+        if self.output_ec_a_type == "volume_ml":
+            self.ec_a_type = 'vol'
+
+        self.ec_b_type = 'sec'
+        if self.output_ec_b_type == "volume_ml":
+            self.ec_b_type = 'vol'
+
+        self.ec_c_type = 'sec'
+        if self.output_ec_c_type == "volume_ml":
+            self.ec_c_type = 'vol'
+
+        self.ec_d_type = 'sec'
+        if self.output_ec_d_type == "volume_ml":
+            self.ec_d_type = 'vol'
 
         if "," in self.email_notification:
             self.email_notification = self.email_notification.split(",")
+
+        if self.output_ec_a_device_id:
+            self.ratio_letters.append("A")
+            self.ratio_numbers.append(self.output_ec_a_amount)
+            self.list_doses.append(f"{self.output_ec_a_amount} {self.output_types[self.output_ec_a_type]} Nut A")
+        if self.output_ec_b_device_id:
+            self.ratio_letters.append("B")
+            self.ratio_numbers.append(self.output_ec_b_amount)
+            self.list_doses.append(f"{self.output_ec_b_amount} {self.output_types[self.output_ec_b_type]} Nut B")
+        if self.output_ec_c_device_id:
+            self.ratio_letters.append("C")
+            self.ratio_numbers.append(self.output_ec_c_amount)
+            self.list_doses.append(f"{self.output_ec_c_amount} {self.output_types[self.output_ec_c_type]} Nut C")
+        if self.output_ec_d_device_id:
+            self.ratio_letters.append("D")
+            self.ratio_numbers.append(self.output_ec_d_amount)
+            self.list_doses.append(f"{self.output_ec_d_amount} {self.output_types[self.output_ec_d_type]} Nut D")
 
     def loop(self):
         if self.timer_loop > time.time():
@@ -510,18 +634,22 @@ class CustomModule(AbstractFunction):
                     self.email(message)
             return
 
+        #
         # First check if pH is dangerously low or high, and adjust if it is
-        if last_measurement_ph[1] < self.danger_range_ph_low:  # pH dangerously low, add base (pH up)
+        #
+
+        # pH dangerously low, add base (pH up)
+        if last_measurement_ph[1] < self.danger_range_ph_low:
             msg = f"pH is dangerously low: {last_measurement_ph[1]}. Should be > {self.danger_range_ph_low}. " \
-                  f"Dispensing {self.output_ph_duration_sec} sec base"
+                  f"Dispensing {self.output_ph_amount} {self.output_types[self.output_ph_type]} base"
             self.logger.debug(msg)
             message += msg
 
             output_on_off = threading.Thread(
                 target=self.control.output_on_off,
                 args=(self.output_ph_raise_device_id, "on",),
-                kwargs={'output_type': 'sec',
-                        'amount': self.output_ph_duration_sec,
+                kwargs={'output_type': self.ph_type,
+                        'amount': self.output_ph_amount,
                         'output_channel': self.output_ph_raise_channel})
             output_on_off.start()
 
@@ -530,17 +658,18 @@ class CustomModule(AbstractFunction):
                     self.email_timers['notify_ph'] = time.time() + (self.email_timer_duration_hours * 60 * 60)
                     self.email(message)
 
-        elif last_measurement_ph[1] > self.danger_range_ph_high:  # pH dangerously high, add acid (pH down)
+        # pH dangerously high, add acid (pH down)
+        elif last_measurement_ph[1] > self.danger_range_ph_high:
             msg = f"pH is dangerously high: {last_measurement_ph[1]}. Should be < {self.danger_range_ph_high}. " \
-                  f"Dispensing {self.output_ph_duration_sec} sec acid"
+                  f"Dispensing {self.output_ph_amount} {self.output_types[self.output_ph_type]} acid"
             self.logger.debug(msg)
             message += msg
 
             output_on_off = threading.Thread(
                 target=self.control.output_on_off,
                 args=(self.output_ph_lower_device_id, "on",),
-                kwargs={'output_type': 'sec',
-                        'amount': self.output_ph_duration_sec,
+                kwargs={'output_type': self.ph_type,
+                        'amount': self.output_ph_amount,
                         'output_channel': self.output_ph_lower_channel})
             output_on_off.start()
 
@@ -549,58 +678,54 @@ class CustomModule(AbstractFunction):
                     self.email_timers['notify_ph'] = time.time() + (self.email_timer_duration_hours * 60 * 60)
                     self.email(message)
 
+        #
         # If pH isn't dangerously low or high, check if EC is within range
-        elif last_measurement_ec[1] < self.range_ec[0]:  # EC too low, add nutrient
-            list_doses = []
-            if self.output_ec_nutrient_a_device_id:
-                list_doses.append(f"{self.output_ec_a_duration_sec} sec Nut A")
-            if self.output_ec_nutrient_b_device_id:
-                list_doses.append(f"{self.output_ec_b_duration_sec} sec Nut B")
-            if self.output_ec_nutrient_c_device_id:
-                list_doses.append(f"{self.output_ec_c_duration_sec} sec Nut C")
-            if self.output_ec_nutrient_d_device_id:
-                list_doses.append(f"{self.output_ec_d_duration_sec} sec Nut D")
+        #
 
-            self.logger.debug(f"EC: {last_measurement_ec[1]}. Should be > {self.range_ec[0]}. "
-                              f"Dosing: {', '.join(list_doses)}")
+        # EC too low, add nutrient
+        elif last_measurement_ec[1] < self.range_ec[0]:
+            self.logger.debug(
+                f"EC: {last_measurement_ec[1]}. Should be > {self.range_ec[0]}. "
+                f"Dosing {':'.join(self.ratio_numbers)} ({':'.join(map(str, self.ratio_letters))}): {', '.join(self.list_doses)})")
 
-            if self.output_ec_nutrient_a_device_id:
+            if self.output_ec_a_device_id:
                 output_on_off = threading.Thread(
                     target=self.control.output_on_off,
-                    args=(self.output_ec_nutrient_a_device_id, "on",),
-                    kwargs={'output_type': 'sec',
-                            'amount': self.output_ec_a_duration_sec,
+                    args=(self.output_ec_a_device_id, "on",),
+                    kwargs={'output_type': self.ec_a_type,
+                            'amount': self.output_ec_a_amount,
                             'output_channel': self.output_ec_a_channel})
                 output_on_off.start()
 
-            if self.output_ec_nutrient_b_device_id:
+            if self.output_ec_b_device_id:
                 output_on_off = threading.Thread(
                     target=self.control.output_on_off,
-                    args=(self.output_ec_nutrient_b_device_id, "on",),
-                    kwargs={'output_type': 'sec',
-                            'amount': self.output_ec_b_duration_sec,
+                    args=(self.output_ec_b_device_id, "on",),
+                    kwargs={'output_type': self.ec_b_type,
+                            'amount': self.output_ec_b_amount,
                             'output_channel': self.output_ec_b_channel})
                 output_on_off.start()
 
-            if self.output_ec_nutrient_c_device_id:
+            if self.output_ec_c_device_id:
                 output_on_off = threading.Thread(
                     target=self.control.output_on_off,
-                    args=(self.output_ec_nutrient_c_device_id, "on",),
-                    kwargs={'output_type': 'sec',
-                            'amount': self.output_ec_c_duration_sec,
+                    args=(self.output_ec_c_device_id, "on",),
+                    kwargs={'output_type': self.ec_c_type,
+                            'amount': self.output_ec_c_amount,
                             'output_channel': self.output_ec_c_channel})
                 output_on_off.start()
 
-            if self.output_ec_nutrient_d_device_id:
+            if self.output_ec_d_device_id:
                 output_on_off = threading.Thread(
                     target=self.control.output_on_off,
-                    args=(self.output_ec_nutrient_d_device_id, "on",),
-                    kwargs={'output_type': 'sec',
-                            'amount': self.output_ec_d_duration_sec,
+                    args=(self.output_ec_d_device_id, "on",),
+                    kwargs={'output_type': self.ec_d_type,
+                            'amount': self.output_ec_d_amount,
                             'output_channel': self.output_ec_d_channel})
                 output_on_off.start()
 
-        elif last_measurement_ec[1] > self.range_ec[1]:  # EC too high, add nutrient
+        # EC too high, add nutrient
+        elif last_measurement_ec[1] > self.range_ec[1]:
             msg = f"EC: {last_measurement_ec[1]}. Should be < {self.range_ec[1]}. Need to add water to dilute!"
             self.logger.debug(msg)
 
@@ -610,30 +735,35 @@ class CustomModule(AbstractFunction):
                     message += msg
                     self.email(message)
 
+        #
         # If pH is in range, make sure pH is within range
-        elif last_measurement_ph[1] < self.range_ph[0]:  # pH too low, add base (pH up)
+        #
+
+        # pH too low, add base (pH up)
+        elif last_measurement_ph[1] < self.range_ph[0]:
             self.logger.debug(
                 f"pH is {last_measurement_ph[1]}. Should be > {self.range_ph[0]}. "
-                f"Dispensing {self.output_ph_duration_sec} sec base")
+                f"Dispensing {self.output_ph_amount} {self.output_types[self.output_ph_type]} base")
 
             output_on_off = threading.Thread(
                 target=self.control.output_on_off,
                 args=(self.output_ph_raise_device_id, "on",),
-                kwargs={'output_type': 'sec',
-                        'amount': self.output_ph_duration_sec,
+                kwargs={'output_type': self.ph_type,
+                        'amount': self.output_ph_amount,
                         'output_channel': self.output_ph_raise_channel})
             output_on_off.start()
 
-        elif last_measurement_ph[1] > self.range_ph[1]:  # pH too high, add acid (pH down)
+        # pH too high, add acid (pH down)
+        elif last_measurement_ph[1] > self.range_ph[1]:
             self.logger.debug(
                 f"pH is {last_measurement_ph[1]}. Should be < {self.range_ph[1]}. "
-                f"Dispensing {self.output_ph_duration_sec} sec acid")
+                f"Dispensing {self.output_ph_amount} {self.output_types[self.output_ph_type]} acid")
 
             output_on_off = threading.Thread(
                 target=self.control.output_on_off,
                 args=(self.output_ph_lower_device_id, "on",),
-                kwargs={'output_type': 'sec',
-                        'amount': self.output_ph_duration_sec,
+                kwargs={'output_type': self.ph_type,
+                        'amount': self.output_ph_amount,
                         'output_channel': self.output_ph_lower_channel})
             output_on_off.start()
 
@@ -656,3 +786,13 @@ class CustomModule(AbstractFunction):
         self.email_timers['notify_ec'] = 0
         self.email_timers['notify_ph'] = 0
         self.email_timers['notify_none'] = 0
+
+    def function_status(self):
+        return_str = {
+            'string_status': f"Regulating"
+                             f"<br>pH: {self.range_ph[0]:.2f} - {self.range_ph[1]:.2f} at {self.output_ph_amount:.2f} {self.output_types[self.output_ph_type]}"
+                             f"<br>EC: {self.range_ec[0]:.2f} - {self.range_ec[1]:.2f} at {':'.join(map(str, self.ratio_numbers))} ({':'.join(self.ratio_letters)})",
+            'error': []
+        }
+        self.logger.info(f"TEST00: {return_str}")
+        return return_str